#25 extract form to EditMenuForm
